+++ classes/ui/wpdk-menu.php
@@ -123,7 +123,7 @@ class WPDKMenu {
   /**
    * Return the WPDK menu info by name of view controller of submenu item
    *
-   * @param string $view_controller The view controller class
+   * @param string $view_controller The view controller class name
    *
    * @return array
    */
@@ -161,6 +161,46 @@ class WPDKMenu {
       }
     }
     return $url;
+  }
+
+  /**
+   * Return the hook id for a view controller
+   *
+   * @brief Hook id
+   * @since 1.2.0
+   *
+   * @param string $view_controller Class name of view controller
+   *
+   * @return string
+   */
+  public static function hook( $view_controller )
+  {
+    $info = self::menu( $view_controller );
+
+    if ( !empty( $info ) ) {
+      return $info['hook'];
+    }
+    return false;
+  }
+
+  /**
+   * Return the page id for a view controller
+   *
+   * @brief Page id
+   * @since 1.2.0
+   *
+   * @param string $view_controller Class name of view controller
+   *
+   * @return string
+   */
+  public static function page( $view_controller )
+  {
+    $info = self::menu( $view_controller );
+
+    if ( !empty( $info ) ) {
+      return $info['page'];
+    }
+    return false;
   }
 
   /**
@@ -356,10 +396,35 @@ class WPDKMenu {
    *
    * @return array
    */
-  public static function addSubMenuAt( &$menus, $menu_item, $index) {
+  public static function addSubMenuAt( &$menus, $menu_item, $index )
+  {
     $key = key( $menus );
-    $menus[$key]['subMenus'] = WPDKArray::insert( $menus[$key]['subMenus'], array( $menu_item ), $index );
+    if ( isset( $menus[$key]['subMenus'] ) ) {
+      $menus[$key]['subMenus'] = WPDKArray::insert( $menus[$key]['subMenus'], array( $menu_item ), $index );
+    }
+    else {
+      $menus[$key] = WPDKArray::insert( $menus[$key], array( $menu_item ), $index );
+    }
     return $menus;
+  }
+
+  /**
+   * Return TRUE if the displayed page is the view controller
+   *
+   * @brief Check displayed page
+   * @since 1.2.0
+   *
+   * @param string $id The menu id
+   *
+   * @return bool
+   */
+  public static function isPageWithMenu( $id )
+  {
+    global $plugin_page;
+    if ( $id === $plugin_page ) {
+      return true;
+    }
+    return false;
   }
 
 }
@@ -463,7 +528,7 @@ class WPDKSubMenu {
     $index  = 1;
     foreach ( $sub_menus as $parent => $sub_menu ) {
       foreach ( $sub_menu as $sub_item ) {
-
+        $item = false;
         if ( is_string( $sub_item ) && WPDKSubMenuDivider::DIVIDER === $sub_item ) {
           $item = new WPDKSubMenuDivider( $parent );
         }
@@ -480,10 +545,13 @@ class WPDKSubMenu {
             }
           }
         }
-        $item->render();
-
+        if( !empty( $item ) ) {
+          $item->render();
+        }
       }
-      $result[] = $item;
+      if( !empty( $item ) ) {
+        $result[] = $item;
+      }
     }
     return $result;
   }
@@ -494,6 +562,8 @@ class WPDKSubMenu {
    * @brief Render
    */
   public function render() {
+
+    global $plugin_page;
 
     $hook = '';
 
@@ -521,6 +591,13 @@ class WPDKSubMenu {
     /* Create the menu item. */
     $this->hookName = add_submenu_page( $this->parent, $this->pageTitle, $menu_title, $this->capability, $this->id, $hook );
 
+    /* Execute this action when the page displayed ids for this submenu view. */
+    if( !empty( $plugin_page ) ) {
+      if( $this->id === $plugin_page ) {
+        do_action( 'wpdk_submenu_page', $this, $plugin_page );
+      }
+    }
+
     if ( !empty( $this->viewController ) && is_string( $this->viewController ) && !function_exists( $this->viewController ) ) {
 
       $GLOBALS[WPDKMenu::GLOBAL_MENU][$this->viewController]['hook']       = $this->hookName;
@@ -537,11 +614,7 @@ class WPDKSubMenu {
 }
 
 /**
- * Description
- *
- * ## Overview
- *
- * Description
+ * Submenu divider
  *
  * @class           WPDKSubMenuDivider
  * @author          =undo= <info@wpxtre.me>